chore: move classes to types
