+++ src/providers/PortfolioProvider/types/positions.types.ts
@@ -1,7 +1,7 @@
 import type { PortfolioPositionsQuery } from '@/app/lib/getPositionsForAddress';
 
-export type { PortfolioDefiPosition } from '../classes/PortfolioDefiPosition';
-export type { PortfolioDeFiPositionsGroup } from '../classes/PortfolioDeFiPositionsGroup';
+export type { PortfolioDefiPosition } from './PortfolioDefiPosition';
+export type { PortfolioDeFiPositionsGroup } from './PortfolioDeFiPositionsGroup';
 
 export type PortfolioPositionsQueryWithoutEvm = Omit<
   PortfolioPositionsQuery,
