+++ src/providers/PortfolioProvider/types/tokens.types.ts
@@ -1,5 +1,5 @@
 import type { Account } from '@lifi/wallet-management';
-import type { PortfolioTokenGroup } from '../classes/PortfolioTokenGroup';
+import type { PortfolioTokenGroup } from './PortfolioTokenGroup';
 
 export type PortfolioAccount = Omit<Account, 'address'> & { address: string };
